+++ python/rsklearn/preprocessing/__init__.py
@@ -1,6 +1,7 @@
 """Preprocessing estimators."""
 
 from ._label_encoder import LabelEncoder
+from ._maxabs_scaler import MaxAbsScaler
 from ._minmax_scaler import MinMaxScaler
 from ._normalizer import Normalizer
 from ._one_hot_encoder import OneHotEncoder
@@ -10,6 +11,7 @@ from ._standard_scaler import StandardScaler
 
 __all__ = [
     "LabelEncoder",
+    "MaxAbsScaler",
     "MinMaxScaler",
     "Normalizer",
     "OneHotEncoder",